--- include/aws/common/error_codes.h
@@ -1,63 +1,0 @@
-#ifndef AWS_COMMON_ERROR_CODES_H
-#define AWS_COMMON_ERROR_CODES_H
-/*
- * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- *  http://aws.amazon.com/apache2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-enum aws_common_error {
-    AWS_ERROR_SUCCESS = 0,
-    AWS_ERROR_OOM,
-    AWS_ERROR_UNKNOWN,
-    AWS_ERROR_SHORT_BUFFER,
-    AWS_ERROR_OVERFLOW_DETECTED,
-    AWS_ERROR_UNSUPPORTED_OPERATION,
-    AWS_ERROR_INVALID_BUFFER_SIZE,
-    AWS_ERROR_INVALID_HEX_STR,
-    AWS_ERROR_INVALID_BASE64_STR,
-    AWS_ERROR_INVALID_INDEX,
-    AWS_ERROR_THREAD_INVALID_SETTINGS,
-    AWS_ERROR_THREAD_INSUFFICIENT_RESOURCE,
-    AWS_ERROR_THREAD_NO_PERMISSIONS,
-    AWS_ERROR_THREAD_NOT_JOINABLE,
-    AWS_ERROR_THREAD_NO_SUCH_THREAD_ID,
-    AWS_ERROR_THREAD_DEADLOCK_DETECTED,
-    AWS_ERROR_MUTEX_NOT_INIT,
-    AWS_ERROR_MUTEX_TIMEOUT,
-    AWS_ERROR_MUTEX_CALLER_NOT_OWNER,
-    AWS_ERROR_MUTEX_FAILED,
-    AWS_ERROR_COND_VARIABLE_INIT_FAILED,
-    AWS_ERROR_COND_VARIABLE_TIMED_OUT,
-    AWS_ERROR_COND_VARIABLE_ERROR_UNKNOWN,
-    AWS_ERROR_CLOCK_FAILURE,
-    AWS_ERROR_LIST_EMPTY,
-    AWS_ERROR_DEST_COPY_TOO_SMALL,
-    AWS_ERROR_LIST_EXCEEDS_MAX_SIZE,
-    AWS_ERROR_LIST_STATIC_MODE_CANT_SHRINK,
-    AWS_ERROR_PRIORITY_QUEUE_FULL,
-    AWS_ERROR_PRIORITY_QUEUE_EMPTY,
-    AWS_ERROR_PRIORITY_QUEUE_BAD_NODE,
-    AWS_ERROR_HASHTBL_ITEM_NOT_FOUND,
-    AWS_ERROR_INVALID_DATE_STR,
-    AWS_ERROR_INVALID_ARGUMENT,
-    AWS_ERROR_RANDOM_GEN_FAILED,
-    AWS_ERROR_MALFORMED_INPUT_STRING,
-    AWS_ERROR_UNIMPLEMENTED,
-    AWS_ERROR_INVALID_STATE,
-    AWS_ERROR_ENVIRONMENT_GET,
-    AWS_ERROR_ENVIRONMENT_SET,
-    AWS_ERROR_ENVIRONMENT_UNSET,
-    AWS_ERROR_END_COMMON_RANGE = 0x03FF
-};
-
-#endif /* AWS_COMMON_ERROR_CODES_H */
